separated DB config file and created login functionality with errors at the moment
