+++ studyingassistant/src/pages/loginPage.html
@@ -7,6 +7,10 @@
     <link href="https://fonts.googleapis.com/css?family=Raleway&display=swap" rel="stylesheet">
     <meta charset='utf-8'>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
+    <!-- The core Firebase JS SDK is always required and must be listed first -->
+    <script src="https://www.gstatic.com/firebasejs/7.13.2/firebase-app.js"></script>
+    <!-- Add additional services that you want to use -->
+    <script src="https://www.gstatic.com/firebasejs/7.13.2/firebase-auth.js"></script>
 </head>
 <body>
     <div class="container loginPage">
@@ -21,7 +25,7 @@
                     <div class="form-group row">
                         <input type="password" class="form-control txtField" id="inputPassword3" placeholder="🔑 Password">
                     </div>
-                    <a href="../pages/mainWindow.html" class="btn btn-outline-primary loginBtn">Submit</a>
+                    <a href="#" class="btn btn-outline-primary loginBtn" type="button" id="submitBtn">Submit</a>
                     <!--for spacing-->
                     <br></br>
                 </form>
@@ -31,9 +35,19 @@
         <p>Don't have an account? <a href="../pages/registerPage.html">Sign Up!</a></p>
     </div>
 
+    <script>
+        require('../scripts/dbConnection.js');
+    </script>
     <!--javascript files for bootstrap-->
+    <script
+        src="https://code.jquery.com/jquery-3.4.1.js"
+        integrity="sha256-WpOohJOqMqqyKL9FccASB9O0KwACQJpFTUBLTYOVvVU="
+        crossorigin="anonymous">
+    </script>
     <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js" integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js" integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous"></script>
+
+    <script src="../scripts/loginPage.js"></script>
 </body>
 </html>
